+++ iocp_chat_server/iocp_chat_server/RedisConfig.h
@@ -2,4 +2,4 @@
 #include <basetsd.h>
 
 const char* REDIS_IP = "127.0.0.1";
-const UINT32 REDIS_PORT = 6379;
+const UINT16 REDIS_PORT = 6379;
